// Rio.Web/src/app/shared/generated/model/account-simple-dto.ts
/**
 * Rio.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class AccountSimpleDto { 
    AccountDisplayName?: string;
    ShortAccountDisplayName?: string;
    AccountID?: number;
    AccountNumber?: number;
    AccountName?: string;
    AccountStatusID?: number;
    Notes?: string;
    UpdateDate?: string;
    AccountVerificationKey?: string;
    AccountVerificationKeyLastUseDate?: string;
    CreateDate?: string;
    InactivateDate?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}